+++ apps/senstate-dashboard/src/app/components/tab-content/watchers-list/watchers-list.component.ts
@@ -1,8 +1,9 @@
-import {Component, OnInit, ChangeDetectionStrategy, Input, TrackByFunction} from '@angular/core';
+import {ChangeDetectionStrategy, Component, Input, OnInit, TrackByFunction} from '@angular/core';
 import {Observable} from "rxjs";
 import {WatcherMeta} from "@senstate/client";
 import {GroupedWatchers, HubService} from "../../../state/hub.service";
 import {DebugToggleService} from "../../../services/debug-toggle.service";
+import {map} from "rxjs/operators";
 
 @Component({
   selector: 'senstate-watchers-list',
@@ -13,9 +14,15 @@ import {DebugToggleService} from "../../../services/debug-toggle.service";
 export class WatchersListComponent implements OnInit {
 
   @Input()
+  public groupedColumnWidth: number = 50;
+
+  @Input()
+  public watcherColumnWidth: number = 50;
+
+  @Input()
   public appId: string;
 
-  public watchers$: Observable<GroupedWatchers[]>;
+  public groupedWatchers$: Observable<GroupedWatchers[]>;
 
   public trackByWatcherFunc: TrackByFunction<WatcherMeta> = (index, item) => {
     return item.watchId;
@@ -29,6 +36,21 @@ export class WatchersListComponent implements OnInit {
               public debugToggle: DebugToggleService) { }
 
   ngOnInit() {
-    this.watchers$ = this.hub.getGroupedWatchersByApp$(this.appId);
+    this.groupedWatchers$ = this.hub.getGroupedWatchersByApp$(this.appId).pipe(
+      map(groupedWatchers => {
+        const hasGroupName = groupedWatchers.some(group => this.hasName(group));
+
+        groupedWatchers.forEach(group => {
+          group.haveGroups = hasGroupName;
+          group.hasName = this.hasName(group);
+        });
+
+        return groupedWatchers;
+      })
+    );
+  }
+
+  private hasName(group: GroupedWatchers) {
+    return group.key && group.key !== 'undefined';
   }
 }