+++ apps/senstate-dashboard/src/app/components/tab-content/watchers-masonry/watchers-masonry.component.spec.ts
@@ -1,8 +1,9 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { WatchersMasonryComponent } from './watchers-masonry.component';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+import {WatchersMasonryComponent} from './watchers-masonry.component';
 import {NO_ERRORS_SCHEMA} from "@angular/core";
 import {HubServiceMock, HubServiceMockProvider} from "@test-utils/mocks";
 import {NEVER} from "rxjs";
+import { DynamicPortalModule, PortalHubService } from "@gewd/ng-utils/dynamic-portal";
 
 describe('WatchersMasonryComponent', () => {
   let component: WatchersMasonryComponent;
@@ -10,9 +11,15 @@ describe('WatchersMasonryComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ WatchersMasonryComponent ],
+      imports: [
+        DynamicPortalModule
+      ],
+      declarations: [
+        WatchersMasonryComponent
+      ],
       providers: [
-        HubServiceMockProvider
+        HubServiceMockProvider,
+        PortalHubService
       ],
       schemas: [NO_ERRORS_SCHEMA]
     })