+++ apps/senstate-dashboard/src/app/components/tab-content/watchers-masonry/watchers-masonry.module.ts
@@ -15,6 +15,8 @@ import {MatCheckboxModule} from "@angular/material/checkbox";
 import {WatchHistoryComponent} from './watchers/watch-history/watch-history.component';
 import {WatchDiffComponent} from "./watchers/watch-diff/watch-diff.component";
 import {SharedModule} from "../../shared/shared.module";
+import {MatFormFieldModule} from "@angular/material/form-field";
+import {DynamicPortalModule} from "@gewd/ng-utils/dynamic-portal";
 
 
 @NgModule({
@@ -41,7 +43,9 @@ import {SharedModule} from "../../shared/shared.module";
         MatIconModule,
         MatButtonModule,
         MatCheckboxModule,
-        SharedModule
+        SharedModule,
+        MatFormFieldModule,
+        DynamicPortalModule
     ]
 })
 export class WatchersMasonryModule {