+++ apps/senstate-dashboard/src/app/directives/css-props/css-props.module.ts
@@ -1,0 +1,15 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {CSSPropsDirective} from "./css-props.directive";
+
+
+@NgModule({
+  declarations: [CSSPropsDirective],
+  exports: [
+    CSSPropsDirective
+  ],
+  imports: [
+    CommonModule
+  ]
+})
+export class CssPropsModule { }